+++ src/main/java/com/example/quizcards/service/ISetFlashcardService.java
@@ -7,11 +7,17 @@ import com.example.quizcards.dto.SetFlashcardUpdateRequest;
 import java.util.List;
 
 public interface ISetFlashcardService {
-    List<IFlashcardDTO> getAllFlashcardBySetId(int setId);
+    List<IFlashcardDTO> getAllFlashcardBySetId(Long setId);
     List<ISetFlashcardDTO> getAll();
-    ISetFlashcardDTO findBySetId(int setId);
-    void addSetFlashcard(String title, String descriptionSet, Boolean isApproved, Boolean isAnonymous, Boolean sharingMode, Long userId, int categoryId);
-    void deleteSetFlashcard(int setId);
+    ISetFlashcardDTO findBySetId(Long setId);
+    void addSetFlashcard(String title,
+                         String descriptionSet,
+                         Boolean isApproved,
+                         Boolean isAnonymous,
+                         Boolean sharingMode,
+                         Long userId,
+                         Long categoryId);
+    void deleteSetFlashcard(Long setId);
     void updateSetFlashcard(SetFlashcardUpdateRequest request);
     List<ISetFlashcardDTO> searchByTitle(String title);
     List<ISetFlashcardDTO> sortByUpdatedDate();
